+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed `InteractiveList` examples
 - Fixed truncation of long text in the `InteractiveList` for IE 11
 - Fixed regular expression for Babel in Webpack
+- Fixed hiding suggestion list of `Autocomplete` after selection for IE 11
 
 ## 0.8.1